pluged in new scheduler

+++ android/expoview/src/main/java/host/exp/exponent/notifications/insecurecheduler/InsecureScheduler.java
@@ -7,7 +7,7 @@ import java.util.HashMap;
 
 public interface InsecureScheduler {
 
-  void schedule(String experienceId, int elapsedTime, int notificationId, HashMap notification, final Context context);
+  void schedule(String experienceId, long elapsedTime, int notificationId, HashMap notification, final Context context);
 
   void cancelScheduled(String experienceId, int notificationId, final Context context);
